Add shuffled word iterator
This can save a good bit of memory for runs with large alphabets
as it no longer requires all of the possible word values to be buffered.
Using an LCG isn't quite as random as a shuffle, but it does guarantee
that we only produce each value exactly once with no memory overhead.